Add files via upload
EKG Coding shows arrhythmia binary coding decisions made by me. Mention of the terms highlighted with green in column D resulted in a coding of 1, blank EKG analysis resulted in NA, and anything else resulted in a coding of 0.

Hypoxemia Analysis produces logistic models for "Arrhythmia" as a function of each of the three Desat LT columns along with ROC curves. No significance was found. This is still found when accounting for various confounding factors such as gender.